requirements.txt
numpy~=1.24.2
tensorflow~=2.11.0
Pillow~=9.4.0
scipy~=1.10.1